NX-24: type surface construction branch modes
Replace raw surface branch mode bytes with a closed enum shared by the
parser and native feature record. Remove the parser's separate whitelist;
typed admission owns the two values. Serde retains numeric modes and rejects
out-of-set values at construction. Other payload branch grammars retain
their independent domains.

Validation: cargo check -q -p cadmpeg-codec-nx --tests; git diff --check.
Hook-Bypass-Reason: User requires the full gate once at mission completion.

+++ crates/cadmpeg-codec-nx/src/om/tests/instances_and_stores.rs
@@ -632,7 +632,7 @@ fn om_surface_feature_branches_require_one_complete_counted_group() {
     assert_eq!(group.family, 0x14);
     assert_eq!(group.header_code, 0x13);
     assert_eq!(group.branches.len(), 2);
-    assert_eq!(group.branches[0].mode, 0x40);
+    assert_eq!(u8::from(group.branches[0].mode), 0x40);
     assert_eq!(group.branches[0].declared_count, 4);
     assert!(group.branches[0].witnessed);
     assert_eq!(group.branches[0].members.len(), 3);
